requirements.txt
replicate>=0.25.0
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0
ffmpeg-python>=0.2.0
openai>=1.12.0
anthropic>=0.18.0
